# target/app.py_openapi.yaml
openapi: 3.0.0
info:
  title: Grocery API
  version: 1.0.0
paths:
  /grocery_items:
    get:
      summary: Get all grocery items
      responses:
        200:
          description: A list of grocery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroceryItem'
    post:
      summary: Add a new grocery item
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroceryItem'
      responses:
        201:
          description: Grocery item added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroceryItem'
  /grocery_items/{item_id}:
    put:
      summary: Update a grocery item
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroceryItem'
      responses:
        200:
          description: Grocery item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroceryItem'
        404:
          description: Item not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete a grocery item
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Grocery item deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
components:
  schemas:
    GroceryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float